+++ src/components/genes/GenesSubscriber.js
@@ -1,95 +1,83 @@
+/* eslint-disable */
 import React, {
   useState, useEffect, useCallback, useMemo,
 } from 'react';
-import PubSub from 'pubsub-js';
-import {
-  EXPRESSION_MATRIX_ADD, CELLS_COLOR, CLEAR_PLEASE_WAIT, RESET,
-} from '../../events';
-import { interpolatePlasma } from '../interpolate-colors';
-import { fromEntries, pluralize } from '../../utils';
+
+import { pluralize } from '../../utils';
+import { useReady } from '../utils';
+import { useCoordination } from '../../app/state/hooks';
+import { componentCoordinationTypes } from '../../app/state/coordination';
+
 import TitleInfo from '../TitleInfo';
 import Genes from './Genes';
 
 export default function GenesSubscriber(props) {
   const {
-    onReady,
+    loaders,
+    coordinationScopes,
     removeGridComponent,
     variablesLabelOverride: variablesLabel = 'gene',
     variablesPluralLabelOverride: variablesPluralLabel = `${variablesLabel}s`,
     theme,
   } = props;
 
-  const [expressionMatrix, setExpressionMatrix] = useState();
-  const [selectedId, setSelectedId] = useState(null);
-  const [urls, setUrls] = useState([]);
+  const [{
+    dataset,
+    geneSelection,
+    geneFilter,
+  }, {
+    setGeneSelection,
+    setGeneFilter,
+    setGeneHighlight,
+  }] = useCoordination(componentCoordinationTypes.genes, coordinationScopes);
 
-  const onReadyCallback = useCallback(onReady, []);
+  const [isReady, setItemIsReady, resetReadyItems] = useReady(
+    ['expression-matrix'],
+  );
+
+  const [geneList, setGeneList] = useState([]);
+
 
   useEffect(() => {
-    const expressionMatrixAddToken = PubSub.subscribe(
-      EXPRESSION_MATRIX_ADD, (msg, { data }) => {
-        const [attrs, arr] = data;
-        setExpressionMatrix({
-          cols: attrs.cols,
-          rows: attrs.rows,
-          matrix: arr.data,
-        });
-      },
-    );
-    const resetToken = PubSub.subscribe(RESET, () => {
-      setUrls([]);
-      setExpressionMatrix(null);
-      setSelectedId({});
-    });
-    onReadyCallback();
-    return () => {
-      PubSub.unsubscribe(expressionMatrixAddToken);
-      PubSub.unsubscribe(resetToken);
-    };
-  }, [onReadyCallback]);
+    resetReadyItems();
 
-  const setSelectedGene = useCallback((newSelectedId) => {
-    setSelectedId(newSelectedId);
-
-    if (expressionMatrix) {
-      const colI = expressionMatrix.cols.indexOf(newSelectedId);
-      const numCols = expressionMatrix.cols.length;
-      if (colI !== -1) {
-        // Create new cellColors map based on the selected gene.
-        const cellColors = new Map(expressionMatrix.rows.map((cellId, rowI) => {
-          const value = expressionMatrix.matrix[rowI * numCols + colI];
-          const cellColor = interpolatePlasma(value / 255);
-          return [cellId, cellColor];
-        }));
-        PubSub.publish(CELLS_COLOR, cellColors);
-      }
+    if(!loaders[dataset]) {
+      return;
     }
-  }, [expressionMatrix]);
 
-  const genesSelected = useMemo(() => {
-    if (!expressionMatrix) {
-      return null;
+    if (loaders[dataset].loaders['expression-matrix']) {
+      loaders[dataset].loaders['expression-matrix'].load().then(({ data }) => {
+        const [attrs] = data;
+        setGeneList(attrs.cols);
+        setItemIsReady('expression-matrix');
+      });
+    } else {
+      // If no gene expression loader is available, set to null.
+      setGeneList([]);
+      // But do not set to ready,
+      // since expression matrix is not optional for genes component.
+      console.warn('Genes component requires expression-matrix data type.');
     }
-    return fromEntries(expressionMatrix.cols.map(geneId => [geneId, geneId === selectedId]));
-  }, [expressionMatrix, selectedId]);
+  }, [loaders, dataset]);
 
-  const numGenes = expressionMatrix ? expressionMatrix.cols.length : 0;
+  const numGenes = geneList.length;
 
   return (
     <TitleInfo
       title="Expression Levels"
       info={`${numGenes} ${pluralize(variablesLabel, variablesPluralLabel, numGenes)}`}
       isScroll
-      urls={urls}
       theme={theme}
       removeGridComponent={removeGridComponent}
+      isReady={isReady}
     >
       <Genes
-        genesSelected={genesSelected}
-        setSelectedGene={setSelectedGene}
-        clearPleaseWait={
-          layerName => PubSub.publish(CLEAR_PLEASE_WAIT, layerName)
-        }
+        geneList={geneList}
+        geneSelection={geneSelection}
+        geneFilter={geneFilter}
+        setGeneSelection={setGeneSelection}
+        setGeneFilter={setGeneFilter}
+        setGeneHighlight={setGeneHighlight}
       />
     </TitleInfo>
   );